More renames > render > renderer
Now, now everything is consistent! :D
