+++ src/java/growthcraft/cellar/renderer/RenderFruitPresser.java
@@ -1,4 +1,4 @@
-package growthcraft.cellar.render;
+package growthcraft.cellar.renderer;
 
 import growthcraft.cellar.block.BlockFruitPresser;
 import growthcraft.core.Utils;
