+++ src/java/growthcraft/core/block/BlockFenceRope.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.Random;
 
 import growthcraft.core.GrowthCraftCore;
-import growthcraft.core.render.RenderFenceRope;
+import growthcraft.core.renderer.RenderFenceRope;
 
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
